Add AbstractSelect::getFilterOptionsForDcGeneral
This method will always return the options list based on id instead of
alias.
Therefore we do not have problems anymore, when editing data in a
language different from the language the ui is using.
The problem was, when the select was referencing a translated MetaModel
and the configured alias column was also translated, then is was tried
to save an alias value in the language of the UI instead of the
currently active dataset rendering an exception to be thrown as the
select attribute does not accept the value (it does not know about).